Fix Windows x86 pipeline failure: handle NumPy behavior in reference (#4955)
* Fix Windows x86 pipeline failure: handle NumPy behavior in onnx.reference


* use astype(np.int64)


---------

+++ onnx/reference/ops/op_non_zero.py
@@ -8,5 +8,6 @@ from onnx.reference.op_run import OpRun
 
 class NonZero(OpRun):
     def _run(self, x):  # type: ignore
-        res = np.vstack(np.nonzero(x))
+        # Specify np.int64 for Windows x86 machines
+        res = np.vstack(np.nonzero(x)).astype(np.int64)
         return (res,)
